Fix signature for request that have colon in URL
This is a bug fix for when a canonical URL has colons (:). This is
common for Lambda function names, for example.

The solution was borrowed from ExAws:
https://github.com/ex-aws/ex_aws/blob/623478ed321ffc6c07fdd7236a2f0e03f1cbd517/lib/ex_aws/request/url.ex#L108

This commit also organizes the signature code in its own module.

Fixes https://github.com/aws-beam/aws-elixir/issues/77
